update baseline methods
